fix: 별자리 컬럼 수정 및 userId 타입 변경으로 인한 수정 작업 (#33)
* 별자리 엔티티 - 설명 컬럼 삭제
* userId 타입 변경으로 인한 수정 작업

+++ src/main/java/connectingstar/tars/habit/response/RunHabitPutResponse.java
@@ -3,6 +3,7 @@ package connectingstar.tars.habit.response;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import connectingstar.tars.habit.domain.RunHabit;
+import jakarta.persistence.criteria.CriteriaBuilder.In;
 import lombok.Getter;
 
 
@@ -28,7 +29,7 @@ public class RunHabitPutResponse {
     /**
      * 사용자 PK
      */
-    private final Long userId;
+    private final Integer userId;
 
 
     /**
